Improve contributor documentation
Remove the paragraph that specifies which tags a new contributor
should work on.  We are not yet operating at the scale where this
would be necessary.  Let's add rules when we need them.

Other than that, I tried to improve the grammar, mostly by adding
articles.

This addresses #27#discussion_r1280738187